make some moves tables smaller

+++ engine/src/cntflip.rs
@@ -5,13 +5,13 @@
 
    Last modified:   December 25, 1999
 */
-pub fn AnyFlips_compact(board: & crate::src::globals::Board, mut inc : &[i32],
+pub fn AnyFlips_compact(board: & crate::src::globals::Board, mut inc : &[i8],
                                sqnum: i32, color: i32, oppcol: i32) -> i32 {
 
     let sq = sqnum;
     loop  {
         let any_drctnl_flips = {
-            let inc: i32 = inc[0];
+            let inc: i32 = inc[0] as i32;
             let mut pt = sq + (inc);
             if board[pt as usize] == oppcol {
                 pt = pt + (inc);
